requirements.txt
streamlit>=1.28.0
pytesseract>=0.3.10
opencv-python>=4.8.0
pillow>=10.0.0
spacy>=3.7.0
pandas>=2.0.0
numpy>=1.25.0
plotly>=5.15.0
scikit-learn>=1.3.0
python-dateutil>=2.8.0
setuptools>=68.0.0
scipy==1.11.3
scikit-image>=0.21.0
# Day 3 MVP requirements - add to your project requirements
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
joblib>=1.2.0
numpy>=1.23.0
pandas
matplotlib

fastapi
uvicorn[standard]
celery[redis]
boto3
botocore
requests
psycopg2-binary   # chỉ nếu bạn chuyển lên Postgres
# MLOps & monitoring
mlflow>=2.6.0
prometheus_client>=0.16.0
sentry-sdk>=1.21.0
structlog>=23.0.0
